change location of the login & register page
